NO-TICKET - hotfix: make rentEpoch field BigInteger instead of int and long

+++ app/src/main/java/org/p2p/wallet/updates/handler/SolanaAccountUpdateHandler.kt
@@ -47,5 +47,5 @@ private data class BalanceValueResponse(
     @SerializedName("executable")
     val executable: Boolean,
     @SerializedName("rentEpoch")
-    val rentEpoch: Int
+    val rentEpoch: BigInteger
 )
